feat: add improved debugging capabilities (#75 via #115)
* feat: add initial naive line implementation

* test: initial line drawer tests

* feat: define line projector

* feat: add line draw optimization

* feat: simplify line draw expression

* refactor: simplify line draw even more

* test: add another line draw test

* fix: allow arbitrary x coords in lines

* test: add more line drawer tests

tests for steep gradients

* feat: implement line draw steep gradient

* refactor: extract common lerp functionality

makes the code simpler to read but changes the memory complexity
from 0(1) to 0(n). Hope the garbage collector doesn't get too upset
at us.

* misc: increase method visibility

* misc: increase method visibility

* feat: add line drawer demo

* refactor: rename 2d line drawing class

* test: add view transform test

This test explains what really happens when the from vector of the
camera view transform is behind the origin (the camera forward
vector is positive z direction)
The default camera forward vector is in the negative z direction

* feat: add point projection to perspective camera

* feat: define some point projection tests

* fix: update sign of focal distance value

the cameras in the two different books are facing opposite directions
we need to account for this when implementing their perspective
projection equations

* test: use deep array equals in projection tests

* test: fix expected test output for flipped camera

transformed vertices are flipped by the camera transform when the
transform looks in the +ve z direction

* feat: implement aabb line drawer

* refactor: remove line projector abstraction

* feat: add aabb vertex generator

* feat: add recursive local point to world method

allows local points within aabb's to be extracted into world space
for rendering

* test: add test for nested local point to world

* test: add local point to world shape tests

* fix: call correct method in nested shape

* feat: create 3D line draw function

* test: add camera clipping plane distance tests

* test: create camera frustum generation test

* test: create clipping plane generation test

* feat: implement camera clipping plane generation

it feels weird mentally that the plane's vector is a half FOV rotation
of the forward vector, but the maths checks out

* test: add clipping tests for each projection plane

* misc: remove unused import

* feat: implement camera view frustum culling

* docs: improve camera docs

* refactor: extract camera to local transform method

The line processing pipeline requires point transformation to occur
before clipping and clipping occurs before projection.

* docs: update camera method docs

* fix: add out of bounds check to canvas line drawer

The canvas line drawer quietly clamps out of bounds line segments
to be in bounds. This was added as a work around while I diagnose
why the point projection and clipping is acting strange for camera
FOVs that are not equals to 90 degrees

* test: add clip plane gen test for different fov

Tests if the clipping planes are correctly generated when the camera
field of view is 60 degrees instead of 90 degrees

* fix: correct clip plane normal generation

Clip plane generation now rotates the camera forward vector half of
the field of view such that it looks down the plane. The cross product
with local space up,  down, left, right vectors is used to get the vector
that is orthogonal to the rotated forward vector and therefore is
orthogonal
to the clipping plane

* misc: put warn message on single output line

* fix: flip pixel locations in line drawer

Not sure why (someone smarter than me could explain)  but the
projected points end up on the wrong half of the normalized screen
coordinates. This change flips them onto the correct half of the screen

* refactor: upgrade torus to use double

* refactor: add children getter for parent type

allows objects with a reference to a parent shape to look at the shape's children. Parent implementers should keep the returned collection
immutable.

* fix: correct torus AABB generation

The torus is on the xy plane in local space but the aabb was assuming
the torus was on the xz plane.

* feat: add recursive shape aabb drawing

* feat: create demo for showing line drawing

* feat: add control over which aabbs are drawn

* refactor: use new draw line api

* test: create line draw behaviour tests

These tests don't have explicit pass/fail conditions. Rather, you can
step through them in the debugger to verify the program
behaviour is correct

* feat: add depth buffer option to line drawer

If a depth buffer canvas is provided to the line drawer, then lines
behind scene objects will not be drawn

* feat: add depth camera distance calculation mode

the line drawer expects the depth values to be the camera local space
distance from a point to the camera origin

* feat: add canvas line drawer depth buffer overload

canvas line drawer method that considers the depth of each point on
the line segment before it writes a pixel to the canvas.

* feat: add depth buffer to line drawer demo scene

* feat: add line drawer demo to param list

* test: add infinite line behaviour check

currently OK with the drawer simply aborting if the line has infinite
length

* docs: add missing java doc for methods and classes